chicago_traffic/requirements.txt
pandas==0.22.0
sodapy==1.4.6
datalab==1.1.2
google-api-python-client
gunicorn==19.7.1
six==1.10
Flask==0.12.2
pandas-gbq==0.3.1